requirements.txt
# Django Server
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
Pillow>=10.0,<12.0
psycopg2-binary>=2.9,<3.0

# Windows Agent
mss>=9.0,<10.0
requests>=2.31,<3.0
pywin32>=306

# Agent packaging
pyinstaller>=6.0,<7.0

# Utilities
python-dotenv>=1.0,<2.0
